Palette of party colours

Colour associated with each party.

Usage

party_palette

Format

A named vector of length 14:

con_code

Factor Constituency code

constituency_name

Factor Constituency name

region_name

Factor Region name

majority

int Difference in vote count between first and second-placed parties

valid_votes17

int Number of valid votes in a constituency

con

int Number of votes for the Conservative Party

lab

int Number of votes for the Labour Party

ld

int Number of votes for the Liberal Democrats

ruk

int Number of votes for the United Kingdom Independence Party (in other datasets ruk represents the Brexit Party in 2019 election and Reform UK in 2024)

green

int Number of votes for the Green Party

snp

int Number of votes for the Scottish National Party

pc

int Number of votes for Plaid Cymru

dup

int Number of votes for the Democratic Unionist Party

sf

int Number of votes for Sinn Fein

sdlp

int Number of votes for the Social Democratic and Labour Party

uup

int Number of votes for the Ulster Unionist Party

apni

int Number of votes for the Alliance Party of Northern Ireland

other

int Number of votes for others

geometry

sfc_GEOMETRY Constituency boundaries 2017/2019

Source

https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Politics_of_the_United_Kingdom/Index_of_United_Kingdom_political_parties_meta_attributes