Vote count for each party in 2017 as projected to 2024 boundaries

The number of votes obtained by each party in each constituency in the 2017 UK General Election projected to 2024 boundaries.

Usage

votes17proj

Format

A dataframe with 650 rows and 31 variables:

con_code

Factor Constituency code

apni17

int Number of votes for the Alliance Party of Northern Ireland

con17

int Number of votes for the Conservative Party

dup17

int Number of votes for the Democratic Unionist Party

green17

int Number of votes for the Green Party

lab17

int Number of votes for the Labour Party

ld17

int Number of votes for the Liberal Democrats

other17

int Number of votes for others

pc17

int Number of votes for Plaid Cymru

ruk17

int Number of votes for Reform UK (in other datasets ruk represents the Brexit Party in 2019 election and Reform UK in 2024)

sdlp17

int Number of votes for the Social Democratic and Labour Party

sf17

int Number of votes for Sinn Fein

snp17

int Number of votes for the Scottish National Party

uup17

int Number of votes for the Ulster Unionist Party

first_party17

Factor Party with most votes in a constituency

second_party17

Factor Party with second-most votes in a constituency

third_party17

Factor Party with third-most votes in a constituency

valid_votes17

int Number of valid votes in a constituency

apni17_pct

num Percentage of valid votes for the Alliance Party of Northern Ireland

con17_pct

num Percentage of valid votes for the Conservative Party

dup17_pct

num Percentage of valid votes for the Democratic Unionist Party

green17_pct

num Percentage of valid votes for the Green Party

lab17_pct

num Percentage of valid votes for the Labour Party

ld17_pct

num Percentage of valid votes for the Liberal Democrats

other17_pct

num Percentage of valid votes for others

pc17_pct

num Percentage of valid votes for Plaid Cymru

ruk17_pct

num Percentage of valid votes for Reform UK (in other datasets ruk represents the Brexit Party in 2019 election and Reform UK in 2024)

sdlp17_pct

num Percentage of valid votes for the Social Democratic and Labour Party

sf17_pct

num Percentage of valid votes for Sinn Fein

snp17_pct

num Percentage of valid votes for the Scottish National Party

uup17_pct

num Percentage of valid votes for the Ulster Unionist Party

Source

https://researchbriefings.files.parliament.uk/documents/CBP-7979/HoC-GE2017-results-by-constituency.csv https://commonslibrary.parliament.uk/boundary-review-2023-which-seats-will-change/