Vote count for each party in 2017, 2019 and 2024, all as projected to hexagonal representation of 2024 boundaries

The number of votes obtained by each party in each constituency in the 2017, 2019 and 2024 UK General Elections, all projected to hexagonal representation of 2024 boundaries.

Usage

votestogether_hex

Format

An sf dataframe with 650 rows and 99 variables:

con_code

Factor Constituency code

constituency_name

Factor Constituency name

region_name

Factor Region name

first_party24

Factor Party with most votes in a constituency

second_party24

Factor Party with second-most votes in a constituency

third_party24

Factor Party with third-most votes in a constituency

con24

int Number of votes for the Conservative Party

lab24

int Number of votes for the Labour Party

ld24

int Number of votes for the Liberal Democrats

ruk24

int Number of votes for Reform UK (in other datasets ruk represents the United Kingdom Independence Party in 2017 election and the Brexit Party in 2019)

green24

int Number of votes for the Green Party

snp24

int Number of votes for the Scottish National Party

pc24

int Number of votes for Plaid Cymru

dup24

int Number of votes for the Democratic Unionist Party

sf24

int Number of votes for Sinn Fein

sdlp24

int Number of votes for the Social Democratic and Labour Party

uup24

int Number of votes for the Ulster Unionist Party

apni24

int Number of votes for the Alliance Party of Northern Ireland

other24

int Number of votes for others

valid_votes24

int Number of valid votes in a constituency

majority

int Difference in vote count between first and second-placed parties

con24_pct

num Percentage of valid votes for the Conservative Party

lab24_pct

num Percentage of valid votes for the Labour Party

ld24_pct

num Percentage of valid votes for the Liberal Democrats

ruk24_pct

num Percentage of valid votes for Reform UK (in other datasets ruk represents the United Kingdom Independence Party in 2017 election and the Brexit Party in 2019)

green24_pct

num Percentage of valid votes for the Green Party

snp24_pct

num Percentage of valid votes for the Scottish National Party

pc24_pct

num Percentage of valid votes for Plaid Cymru

dup24_pct

num Percentage of valid votes for the Democratic Unionist Party

sf24_pct

num Percentage of valid votes for Sinn Fein

sdlp24_pct

num Percentage of valid votes for the Social Democratic and Labour Party

uup24_pct

num Percentage of valid votes for the Ulster Unionist Party

apni24_pct

num Percentage of valid votes for the Alliance Party of Northern Ireland

other24_pct

num Percentage of valid votes for others

apni19

int Number of votes for the Alliance Party of Northern Ireland

con19

int Number of votes for the Conservative Party

dup19

int Number of votes for the Democratic Unionist Party

green19

int Number of votes for the Green Party

lab19

int Number of votes for the Labour Party

ld19

int Number of votes for the Liberal Democrats

other19

int Number of votes for others

pc19

int Number of votes for Plaid Cymru

ruk19

int Number of votes for Reform UK (in other datasets ruk represents the United Kingdom Independence Party in 2017 election and the Reform UK in 2024)

sdlp19

int Number of votes for the Social Democratic and Labour Party

sf19

int Number of votes for Sinn Fein

snp19

int Number of votes for the Scottish National Party

uup19

int Number of votes for the Ulster Unionist Party

first_party19

Factor Party with most votes in a constituency

second_party19

Factor Party with second-most votes in a constituency

third_party19

Factor Party with third-most votes in a constituency

valid_votes19

int Number of valid votes in a constituency

apni19_pct

num Percentage of valid votes for the Alliance Party of Northern Ireland

con19_pct

num Percentage of valid votes for the Conservative Party

dup19_pct

num Percentage of valid votes for the Democratic Unionist Party

green19_pct

num Percentage of valid votes for the Green Party

lab19_pct

num Percentage of valid votes for the Labour Party

ld19_pct

num Percentage of valid votes for the Liberal Democrats

other19_pct

num Percentage of valid votes for others

pc19_pct

num Percentage of valid votes for Plaid Cymru

ruk19_pct

num Percentage of valid votes for Reform UK (in other datasets ruk represents the United Kingdom Independence Party in 2017 election and the Reform UK in 2024)

sdlp19_pct

num Percentage of valid votes for the Social Democratic and Labour Party

sf19_pct

num Percentage of valid votes for Sinn Fein

snp19_pct

num Percentage of valid votes for the Scottish National Party

uup19_pct

num Percentage of valid votes for the Ulster Unionist Party

con17

int Number of votes for the Conservative Party

dup17

int Number of votes for the Democratic Unionist Party

green17

int

lab17

int Number of votes for the Labour Party

ld17

int Number of votes for the Liberal Democrats

other17

int Number of votes for others

pc17

int Number of votes for Plaid Cymru

ruk17

int Number of votes for Reform UK (in other datasets ruk represents the Brexit Party in 2019 election and Reform UK in 2024)

sdlp17

int Number of votes for the Social Democratic and Labour Party

sf17

int Number of votes for Sinn Fein

snp17

int Number of votes for the Scottish National Party

uup17

int Number of votes for the Ulster Unionist Party

first_party17

Factor Party with most votes in a constituency

second_party17

Factor Party with second-most votes in a constituency

third_party17

Factor Party with third-most votes in a constituency

valid_votes17

int Number of valid votes in a constituency

apni17_pct

num Percentage of valid votes for the Alliance Party of Northern Ireland

con17_pct

num Percentage of valid votes for the Conservative Party

dup17_pct

num Percentage of valid votes for the Democratic Unionist Party

green17_pct

num Percentage of valid votes for the Green Party

lab17_pct

num Percentage of valid votes for the Labour Party

ld17_pct

num Percentage of valid votes for the Liberal Democrats

other17_pct

num Percentage of valid votes for others

pc17_pct

num Percentage of valid votes for Plaid Cymru

ruk17_pct

num Percentage of valid votes for Reform UK (in other datasets ruk represents the Brexit Party in 2019 election and Reform UK in 2024)

sdlp17_pct

num Percentage of valid votes for the Social Democratic and Labour Party

sf17_pct

num Percentage of valid votes for Sinn Fein

snp17_pct

num Percentage of valid votes for the Scottish National Party

uup17_pct

num Percentage of valid votes for the Ulster Unionist Party

over65

num Proportion of population in constituency aged over 65

degree

num Proportion of population in constituency with a degree

notgoodhealth

num Proportion of population in constituency reporting fair, bad or very bad health

white

num Proportion of population in constituency of white ethnicity

con_name

Factor Constituency name

con_abb

Factor 3-letter abbreviation of constituency name

reg_name

Factor Region name

geometry

sfc_GEOMETRY Hexagonal representation of constituency boundaries 2024

Source

https://researchbriefings.files.parliament.uk/documents/CBP-7979/HoC-GE2017-results-by-constituency.csv https://researchbriefings.files.parliament.uk/documents/CBP-8749/HoC-GE2019-results-by-constituency.csv https://researchbriefings.files.parliament.uk/documents/CBP-10009/HoC-GE2024-results-by-constituency.csv https://statistics.ukdataservice.ac.uk/dataset/?q=census+2021&sort=score+desc%2C+metadata_modified+desc&vocab_Area_type=Westminster+Parliamentary+Constituencies https://commonslibrary.parliament.uk/boundary-review-2023-which-seats-will-change/ Philip Brown and Alasdair Rae, Automatic Knowledge